Refactor Define CustomFinance in modeles/finance.ts
